LayoutTests/fast/dom/prototype-inheritance-expected.txt

Make sure prototypes are set up using the window a property came from, instead of the lexical global object.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS inner.Attr.isInner is true
PASS inner.Attr.constructor.isInner is true
PASS inner.Audio.isInner is true
PASS inner.Audio.constructor.isInner is true
PASS inner.BeforeLoadEvent.isInner is true
PASS inner.BeforeLoadEvent.constructor.isInner is true
PASS inner.CDATASection.isInner is true
PASS inner.CDATASection.constructor.isInner is true
PASS inner.CSSCharsetRule.isInner is true
PASS inner.CSSCharsetRule.constructor.isInner is true
PASS inner.CSSFontFaceRule.isInner is true
PASS inner.CSSFontFaceRule.constructor.isInner is true
PASS inner.CSSImportRule.isInner is true
PASS inner.CSSImportRule.constructor.isInner is true
PASS inner.CSSMediaRule.isInner is true
PASS inner.CSSMediaRule.constructor.isInner is true
PASS inner.CSSPageRule.isInner is true
PASS inner.CSSPageRule.constructor.isInner is true
PASS inner.CSSPrimitiveValue.isInner is true
PASS inner.CSSPrimitiveValue.constructor.isInner is true
PASS inner.CSSRule.isInner is true
PASS inner.CSSRule.constructor.isInner is true
PASS inner.CSSRuleList.isInner is true
PASS inner.CSSRuleList.constructor.isInner is true
PASS inner.CSSStyleDeclaration.isInner is true
PASS inner.CSSStyleDeclaration.constructor.isInner is true
PASS inner.CSSStyleRule.isInner is true
PASS inner.CSSStyleRule.constructor.isInner is true
PASS inner.CSSStyleSheet.isInner is true
PASS inner.CSSStyleSheet.constructor.isInner is true
PASS inner.CSSValue.isInner is true
PASS inner.CSSValue.constructor.isInner is true
PASS inner.CSSValueList.isInner is true
PASS inner.CSSValueList.constructor.isInner is true
PASS inner.CSSVariablesDeclaration.isInner is true
PASS inner.CSSVariablesDeclaration.constructor.isInner is true
PASS inner.CSSVariablesRule.isInner is true
PASS inner.CSSVariablesRule.constructor.isInner is true
PASS inner.CanvasRenderingContext2D.isInner is true
PASS inner.CanvasRenderingContext2D.constructor.isInner is true
PASS inner.CharacterData.isInner is true
PASS inner.CharacterData.constructor.isInner is true
PASS inner.ClientRect.isInner is true
PASS inner.ClientRect.constructor.isInner is true
PASS inner.ClientRectList.isInner is true
PASS inner.ClientRectList.constructor.isInner is true
PASS inner.Clipboard.isInner is true
PASS inner.Clipboard.constructor.isInner is true
PASS inner.Comment.isInner is true
PASS inner.Comment.constructor.isInner is true
PASS inner.Counter.isInner is true
PASS inner.Counter.constructor.isInner is true
PASS inner.DOMException.isInner is true
PASS inner.DOMException.constructor.isInner is true
PASS inner.DOMImplementation.isInner is true
PASS inner.DOMImplementation.constructor.isInner is true
PASS inner.DOMParser.isInner is true
PASS inner.DOMParser.constructor.isInner is true
PASS inner.Document.isInner is true
PASS inner.Document.constructor.isInner is true
PASS inner.DocumentFragment.isInner is true
PASS inner.DocumentFragment.constructor.isInner is true
PASS inner.DocumentType.isInner is true
PASS inner.DocumentType.constructor.isInner is true
PASS inner.Element.isInner is true
PASS inner.Element.constructor.isInner is true
PASS inner.Entity.isInner is true
PASS inner.Entity.constructor.isInner is true
PASS inner.EntityReference.isInner is true
PASS inner.EntityReference.constructor.isInner is true
PASS inner.EvalError.isInner is true
PASS inner.EvalError.constructor.isInner is true
PASS inner.Event.isInner is true
PASS inner.Event.constructor.isInner is true
PASS inner.EventException.isInner is true
PASS inner.EventException.constructor.isInner is true
PASS inner.EventSource.isInner is true
PASS inner.EventSource.constructor.isInner is true
PASS inner.File.isInner is true
PASS inner.File.constructor.isInner is true
PASS inner.FileList.isInner is true
PASS inner.FileList.constructor.isInner is true
PASS inner.HTMLAnchorElement.isInner is true
PASS inner.HTMLAnchorElement.constructor.isInner is true
PASS inner.HTMLAppletElement.isInner is true
PASS inner.HTMLAppletElement.constructor.isInner is true
PASS inner.HTMLAreaElement.isInner is true
PASS inner.HTMLAreaElement.constructor.isInner is true
PASS inner.HTMLAudioElement.isInner is true
PASS inner.HTMLAudioElement.constructor.isInner is true
PASS inner.HTMLBRElement.isInner is true
PASS inner.HTMLBRElement.constructor.isInner is true
PASS inner.HTMLBaseElement.isInner is true
PASS inner.HTMLBaseElement.constructor.isInner is true
PASS inner.HTMLBaseFontElement.isInner is true
PASS inner.HTMLBaseFontElement.constructor.isInner is true
PASS inner.HTMLBlockquoteElement.isInner is true
PASS inner.HTMLBlockquoteElement.constructor.isInner is true
PASS inner.HTMLBodyElement.isInner is true
PASS inner.HTMLBodyElement.constructor.isInner is true
PASS inner.HTMLButtonElement.isInner is true
PASS inner.HTMLButtonElement.constructor.isInner is true
PASS inner.HTMLCanvasElement.isInner is true
PASS inner.HTMLCanvasElement.constructor.isInner is true
PASS inner.HTMLCollection.isInner is true
PASS inner.HTMLCollection.constructor.isInner is true
PASS inner.HTMLDListElement.isInner is true
PASS inner.HTMLDListElement.constructor.isInner is true
PASS inner.HTMLDataGridCellElement.isInner is true
PASS inner.HTMLDataGridCellElement.constructor.isInner is true
PASS inner.HTMLDataGridColElement.isInner is true
PASS inner.HTMLDataGridColElement.constructor.isInner is true
PASS inner.HTMLDataGridElement.isInner is true
PASS inner.HTMLDataGridElement.constructor.isInner is true
PASS inner.HTMLDirectoryElement.isInner is true
PASS inner.HTMLDirectoryElement.constructor.isInner is true
PASS inner.HTMLDivElement.isInner is true
PASS inner.HTMLDivElement.constructor.isInner is true
PASS inner.HTMLDocument.isInner is true
PASS inner.HTMLDocument.constructor.isInner is true
PASS inner.HTMLElement.isInner is true
PASS inner.HTMLElement.constructor.isInner is true
PASS inner.HTMLEmbedElement.isInner is true
PASS inner.HTMLEmbedElement.constructor.isInner is true
PASS inner.HTMLFieldSetElement.isInner is true
PASS inner.HTMLFieldSetElement.constructor.isInner is true
PASS inner.HTMLFontElement.isInner is true
PASS inner.HTMLFontElement.constructor.isInner is true
PASS inner.HTMLFormElement.isInner is true
PASS inner.HTMLFormElement.constructor.isInner is true
PASS inner.HTMLFrameElement.isInner is true
PASS inner.HTMLFrameElement.constructor.isInner is true
PASS inner.HTMLFrameSetElement.isInner is true
PASS inner.HTMLFrameSetElement.constructor.isInner is true
PASS inner.HTMLHRElement.isInner is true
PASS inner.HTMLHRElement.constructor.isInner is true
PASS inner.HTMLHeadElement.isInner is true
PASS inner.HTMLHeadElement.constructor.isInner is true
PASS inner.HTMLHeadingElement.isInner is true
PASS inner.HTMLHeadingElement.constructor.isInner is true
PASS inner.HTMLHtmlElement.isInner is true
PASS inner.HTMLHtmlElement.constructor.isInner is true
PASS inner.HTMLIFrameElement.isInner is true
PASS inner.HTMLIFrameElement.constructor.isInner is true
PASS inner.HTMLImageElement.isInner is true
PASS inner.HTMLImageElement.constructor.isInner is true
PASS inner.HTMLInputElement.isInner is true
PASS inner.HTMLInputElement.constructor.isInner is true
PASS inner.HTMLIsIndexElement.isInner is true
PASS inner.HTMLIsIndexElement.constructor.isInner is true
PASS inner.HTMLLIElement.isInner is true
PASS inner.HTMLLIElement.constructor.isInner is true
PASS inner.HTMLLabelElement.isInner is true
PASS inner.HTMLLabelElement.constructor.isInner is true
PASS inner.HTMLLegendElement.isInner is true
PASS inner.HTMLLegendElement.constructor.isInner is true
PASS inner.HTMLLinkElement.isInner is true
PASS inner.HTMLLinkElement.constructor.isInner is true
PASS inner.HTMLMapElement.isInner is true
PASS inner.HTMLMapElement.constructor.isInner is true
PASS inner.HTMLMarqueeElement.isInner is true
PASS inner.HTMLMarqueeElement.constructor.isInner is true
PASS inner.HTMLMediaElement.isInner is true
PASS inner.HTMLMediaElement.constructor.isInner is true
PASS inner.HTMLMenuElement.isInner is true
PASS inner.HTMLMenuElement.constructor.isInner is true
PASS inner.HTMLMetaElement.isInner is true
PASS inner.HTMLMetaElement.constructor.isInner is true
PASS inner.HTMLModElement.isInner is true
PASS inner.HTMLModElement.constructor.isInner is true
PASS inner.HTMLOListElement.isInner is true
PASS inner.HTMLOListElement.constructor.isInner is true
PASS inner.HTMLObjectElement.isInner is true
PASS inner.HTMLObjectElement.constructor.isInner is true
PASS inner.HTMLOptGroupElement.isInner is true
PASS inner.HTMLOptGroupElement.constructor.isInner is true
PASS inner.HTMLOptionElement.isInner is true
PASS inner.HTMLOptionElement.constructor.isInner is true
PASS inner.HTMLParagraphElement.isInner is true
PASS inner.HTMLParagraphElement.constructor.isInner is true
PASS inner.HTMLParamElement.isInner is true
PASS inner.HTMLParamElement.constructor.isInner is true
PASS inner.HTMLPreElement.isInner is true
PASS inner.HTMLPreElement.constructor.isInner is true
PASS inner.HTMLQuoteElement.isInner is true
PASS inner.HTMLQuoteElement.constructor.isInner is true
PASS inner.HTMLScriptElement.isInner is true
PASS inner.HTMLScriptElement.constructor.isInner is true
PASS inner.HTMLSelectElement.isInner is true
PASS inner.HTMLSelectElement.constructor.isInner is true
PASS inner.HTMLStyleElement.isInner is true
PASS inner.HTMLStyleElement.constructor.isInner is true
PASS inner.HTMLTableCaptionElement.isInner is true
PASS inner.HTMLTableCaptionElement.constructor.isInner is true
PASS inner.HTMLTableCellElement.isInner is true
PASS inner.HTMLTableCellElement.constructor.isInner is true
PASS inner.HTMLTableColElement.isInner is true
PASS inner.HTMLTableColElement.constructor.isInner is true
PASS inner.HTMLTableElement.isInner is true
PASS inner.HTMLTableElement.constructor.isInner is true
PASS inner.HTMLTableRowElement.isInner is true
PASS inner.HTMLTableRowElement.constructor.isInner is true
PASS inner.HTMLTableSectionElement.isInner is true
PASS inner.HTMLTableSectionElement.constructor.isInner is true
PASS inner.HTMLTextAreaElement.isInner is true
PASS inner.HTMLTextAreaElement.constructor.isInner is true
PASS inner.HTMLTitleElement.isInner is true
PASS inner.HTMLTitleElement.constructor.isInner is true
PASS inner.HTMLUListElement.isInner is true
PASS inner.HTMLUListElement.constructor.isInner is true
PASS inner.HTMLVideoElement.isInner is true
PASS inner.HTMLVideoElement.constructor.isInner is true
PASS inner.Image.isInner is true
PASS inner.Image.constructor.isInner is true
PASS inner.KeyboardEvent.isInner is true
PASS inner.KeyboardEvent.constructor.isInner is true
PASS inner.MediaError.isInner is true
PASS inner.MediaError.constructor.isInner is true
PASS inner.MediaList.isInner is true
PASS inner.MediaList.constructor.isInner is true
PASS inner.MessageChannel.isInner is true
PASS inner.MessageChannel.constructor.isInner is true
PASS inner.MessageEvent.isInner is true
PASS inner.MessageEvent.constructor.isInner is true
PASS inner.MessagePort.isInner is true
PASS inner.MessagePort.constructor.isInner is true
PASS inner.MimeType.isInner is true
PASS inner.MimeType.constructor.isInner is true
PASS inner.MimeTypeArray.isInner is true
PASS inner.MimeTypeArray.constructor.isInner is true
PASS inner.MouseEvent.isInner is true
PASS inner.MouseEvent.constructor.isInner is true
PASS inner.MutationEvent.isInner is true
PASS inner.MutationEvent.constructor.isInner is true
PASS inner.NamedNodeMap.isInner is true
PASS inner.NamedNodeMap.constructor.isInner is true
PASS inner.Node.isInner is true
PASS inner.Node.constructor.isInner is true
PASS inner.NodeFilter.isInner is true
PASS inner.NodeFilter.constructor.isInner is true
PASS inner.NodeList.isInner is true
PASS inner.NodeList.constructor.isInner is true
PASS inner.Notation.isInner is true
PASS inner.Notation.constructor.isInner is true
PASS inner.Option.isInner is true
PASS inner.Option.constructor.isInner is true
PASS inner.OverflowEvent.isInner is true
PASS inner.OverflowEvent.constructor.isInner is true
PASS inner.PageTransitionEvent.isInner is true
PASS inner.PageTransitionEvent.constructor.isInner is true
PASS inner.Plugin.isInner is true
PASS inner.Plugin.constructor.isInner is true
PASS inner.PluginArray.isInner is true
PASS inner.PluginArray.constructor.isInner is true
PASS inner.ProcessingInstruction.isInner is true
PASS inner.ProcessingInstruction.constructor.isInner is true
PASS inner.ProgressEvent.isInner is true
PASS inner.ProgressEvent.constructor.isInner is true
PASS inner.RGBColor.isInner is true
PASS inner.RGBColor.constructor.isInner is true
PASS inner.Range.isInner is true
PASS inner.Range.constructor.isInner is true
PASS inner.RangeError.isInner is true
PASS inner.RangeError.constructor.isInner is true
PASS inner.RangeException.isInner is true
PASS inner.RangeException.constructor.isInner is true
PASS inner.Rect.isInner is true
PASS inner.Rect.constructor.isInner is true
PASS inner.ReferenceError.isInner is true
PASS inner.ReferenceError.constructor.isInner is true
PASS inner.SVGAngle.isInner is true
PASS inner.SVGAngle.constructor.isInner is true
PASS inner.SVGColor.isInner is true
PASS inner.SVGColor.constructor.isInner is true
PASS inner.SVGException.isInner is true
PASS inner.SVGException.constructor.isInner is true
PASS inner.SVGGradientElement.isInner is true
PASS inner.SVGGradientElement.constructor.isInner is true
PASS inner.SVGLength.isInner is true
PASS inner.SVGLength.constructor.isInner is true
PASS inner.SVGMarkerElement.isInner is true
PASS inner.SVGMarkerElement.constructor.isInner is true
PASS inner.SVGPaint.isInner is true
PASS inner.SVGPaint.constructor.isInner is true
PASS inner.SVGPathSeg.isInner is true
PASS inner.SVGPathSeg.constructor.isInner is true
PASS inner.SVGPreserveAspectRatio.isInner is true
PASS inner.SVGPreserveAspectRatio.constructor.isInner is true
PASS inner.SVGRenderingIntent.isInner is true
PASS inner.SVGRenderingIntent.constructor.isInner is true
PASS inner.SVGTextContentElement.isInner is true
PASS inner.SVGTextContentElement.constructor.isInner is true
PASS inner.SVGTextPathElement.isInner is true
PASS inner.SVGTextPathElement.constructor.isInner is true
PASS inner.SVGTransform.isInner is true
PASS inner.SVGTransform.constructor.isInner is true
PASS inner.SVGUnitTypes.isInner is true
PASS inner.SVGUnitTypes.constructor.isInner is true
PASS inner.SharedWorker.isInner is true
PASS inner.SharedWorker.constructor.isInner is true
PASS inner.Storage.isInner is true
PASS inner.Storage.constructor.isInner is true
PASS inner.StorageEvent.isInner is true
PASS inner.StorageEvent.constructor.isInner is true
PASS inner.StyleSheet.isInner is true
PASS inner.StyleSheet.constructor.isInner is true
PASS inner.StyleSheetList.isInner is true
PASS inner.StyleSheetList.constructor.isInner is true
PASS inner.SyntaxError.isInner is true
PASS inner.SyntaxError.constructor.isInner is true
PASS inner.Text.isInner is true
PASS inner.Text.constructor.isInner is true
PASS inner.TextEvent.isInner is true
PASS inner.TextEvent.constructor.isInner is true
PASS inner.TextMetrics.isInner is true
PASS inner.TextMetrics.constructor.isInner is true
PASS inner.TypeError.isInner is true
PASS inner.TypeError.constructor.isInner is true
PASS inner.UIEvent.isInner is true
PASS inner.UIEvent.constructor.isInner is true
PASS inner.URIError.isInner is true
PASS inner.URIError.constructor.isInner is true
PASS inner.WebKitAnimationEvent.isInner is true
PASS inner.WebKitAnimationEvent.constructor.isInner is true
PASS inner.WebKitCSSKeyframeRule.isInner is true
PASS inner.WebKitCSSKeyframeRule.constructor.isInner is true
PASS inner.WebKitCSSKeyframesRule.isInner is true
PASS inner.WebKitCSSKeyframesRule.constructor.isInner is true
PASS inner.WebKitCSSMatrix.isInner is true
PASS inner.WebKitCSSMatrix.constructor.isInner is true
PASS inner.WebKitCSSTransformValue.isInner is true
PASS inner.WebKitCSSTransformValue.constructor.isInner is true
PASS inner.WebKitPoint.isInner is true
PASS inner.WebKitPoint.constructor.isInner is true
PASS inner.WebKitTransitionEvent.isInner is true
PASS inner.WebKitTransitionEvent.constructor.isInner is true
PASS inner.WebSocket.isInner is true
PASS inner.WebSocket.constructor.isInner is true
PASS inner.WheelEvent.isInner is true
PASS inner.WheelEvent.constructor.isInner is true
PASS inner.Worker.isInner is true
PASS inner.Worker.constructor.isInner is true
PASS inner.XMLDocument.isInner is true
PASS inner.XMLDocument.constructor.isInner is true
PASS inner.XMLHttpRequest.isInner is true
PASS inner.XMLHttpRequest.constructor.isInner is true
PASS inner.XMLHttpRequestException.isInner is true
PASS inner.XMLHttpRequestException.constructor.isInner is true
PASS inner.XMLHttpRequestUpload.isInner is true
PASS inner.XMLHttpRequestUpload.constructor.isInner is true
PASS inner.XMLSerializer.isInner is true
PASS inner.XMLSerializer.constructor.isInner is true
PASS inner.XPathEvaluator.isInner is true
PASS inner.XPathEvaluator.constructor.isInner is true
PASS inner.XPathException.isInner is true
PASS inner.XPathException.constructor.isInner is true
PASS inner.XPathResult.isInner is true
PASS inner.XPathResult.constructor.isInner is true
PASS inner.XSLTProcessor.isInner is true
PASS inner.XSLTProcessor.constructor.isInner is true
FAIL inner.addEventListener.isInner should be true. Was false.
FAIL inner.addEventListener.constructor.isInner should be true. Was false.
FAIL inner.alert.isInner should be true. Was false.
FAIL inner.alert.constructor.isInner should be true. Was false.
PASS inner.applicationCache.isInner is true
PASS inner.applicationCache.constructor.isInner is true
FAIL inner.atob.isInner should be true. Was false.
FAIL inner.atob.constructor.isInner should be true. Was false.
FAIL inner.blur.isInner should be true. Was false.
FAIL inner.blur.constructor.isInner should be true. Was false.
FAIL inner.btoa.isInner should be true. Was false.
FAIL inner.btoa.constructor.isInner should be true. Was false.
FAIL inner.captureEvents.isInner should be true. Was false.
FAIL inner.captureEvents.constructor.isInner should be true. Was false.
FAIL inner.clearInterval.isInner should be true. Was false.
FAIL inner.clearInterval.constructor.isInner should be true. Was false.
FAIL inner.clearTimeout.isInner should be true. Was false.
FAIL inner.clearTimeout.constructor.isInner should be true. Was false.
PASS inner.clientInformation.isInner is true
PASS inner.clientInformation.constructor.isInner is true
FAIL inner.close.isInner should be true. Was false.
FAIL inner.close.constructor.isInner should be true. Was false.
FAIL inner.closed.isInner should be true. Was false.
FAIL inner.closed.constructor.isInner should be true. Was false.
FAIL inner.confirm.isInner should be true. Was false.
FAIL inner.confirm.constructor.isInner should be true. Was false.
PASS inner.console.isInner is true
PASS inner.console.constructor.isInner is true
FAIL inner.defaultStatus.isInner should be true. Was false.
FAIL inner.defaultStatus.constructor.isInner should be true. Was false.
FAIL inner.defaultstatus.isInner should be true. Was false.
FAIL inner.defaultstatus.constructor.isInner should be true. Was false.
FAIL inner.devicePixelRatio.isInner should be true. Was false.
FAIL inner.devicePixelRatio.constructor.isInner should be true. Was false.
FAIL inner.dispatchEvent.isInner should be true. Was false.
FAIL inner.dispatchEvent.constructor.isInner should be true. Was false.
PASS inner.document.isInner is true
PASS inner.document.constructor.isInner is true
FAIL inner.find.isInner should be true. Was false.
FAIL inner.find.constructor.isInner should be true. Was false.
FAIL inner.focus.isInner should be true. Was false.
FAIL inner.focus.constructor.isInner should be true. Was false.
FAIL inner.frameElement.isInner should be true. Was false.
FAIL inner.frameElement.constructor.isInner should be true. Was false.
PASS inner.frames.isInner is true
PASS inner.frames.constructor.isInner is true
FAIL inner.getComputedStyle.isInner should be true. Was false.
FAIL inner.getComputedStyle.constructor.isInner should be true. Was false.
FAIL inner.getMatchedCSSRules.isInner should be true. Was false.
FAIL inner.getMatchedCSSRules.constructor.isInner should be true. Was false.
FAIL inner.getSelection.isInner should be true. Was false.
FAIL inner.getSelection.constructor.isInner should be true. Was false.
PASS inner.history.isInner is true
PASS inner.history.constructor.isInner is true
FAIL inner.innerHeight.isInner should be true. Was false.
FAIL inner.innerHeight.constructor.isInner should be true. Was false.
FAIL inner.innerWidth.isInner should be true. Was false.
FAIL inner.innerWidth.constructor.isInner should be true. Was false.
FAIL inner.isInner.isInner should be true. Was false.
FAIL inner.isInner.constructor.isInner should be true. Was false.
FAIL inner.length.isInner should be true. Was false.
FAIL inner.length.constructor.isInner should be true. Was false.
PASS inner.localStorage.isInner is true
PASS inner.localStorage.constructor.isInner is true
PASS inner.location.isInner is true
PASS inner.location.constructor.isInner is true
PASS inner.locationbar.isInner is true
PASS inner.locationbar.constructor.isInner is true
PASS inner.media.isInner is true
PASS inner.media.constructor.isInner is true
PASS inner.menubar.isInner is true
PASS inner.menubar.constructor.isInner is true
FAIL inner.moveBy.isInner should be true. Was false.
FAIL inner.moveBy.constructor.isInner should be true. Was false.
FAIL inner.moveTo.isInner should be true. Was false.
FAIL inner.moveTo.constructor.isInner should be true. Was false.
FAIL inner.name.isInner should be true. Was false.
FAIL inner.name.constructor.isInner should be true. Was false.
PASS inner.navigator.isInner is true
PASS inner.navigator.constructor.isInner is true
FAIL inner.offscreenBuffering.isInner should be true. Was false.
FAIL inner.offscreenBuffering.constructor.isInner should be true. Was false.
FAIL inner.open.isInner should be true. Was false.
FAIL inner.open.constructor.isInner should be true. Was false.
FAIL inner.openDatabase.isInner should be true. Was false.
FAIL inner.openDatabase.constructor.isInner should be true. Was false.
FAIL inner.outerHeight.isInner should be true. Was false.
FAIL inner.outerHeight.constructor.isInner should be true. Was false.
FAIL inner.outerWidth.isInner should be true. Was false.
FAIL inner.outerWidth.constructor.isInner should be true. Was false.
FAIL inner.pageXOffset.isInner should be true. Was false.
FAIL inner.pageXOffset.constructor.isInner should be true. Was false.
FAIL inner.pageYOffset.isInner should be true. Was false.
FAIL inner.pageYOffset.constructor.isInner should be true. Was false.
PASS inner.personalbar.isInner is true
PASS inner.personalbar.constructor.isInner is true
FAIL inner.postMessage.isInner should be true. Was false.
FAIL inner.postMessage.constructor.isInner should be true. Was false.
FAIL inner.print.isInner should be true. Was false.
FAIL inner.print.constructor.isInner should be true. Was false.
FAIL inner.prompt.isInner should be true. Was false.
FAIL inner.prompt.constructor.isInner should be true. Was false.
FAIL inner.releaseEvents.isInner should be true. Was false.
FAIL inner.releaseEvents.constructor.isInner should be true. Was false.
FAIL inner.removeEventListener.isInner should be true. Was false.
FAIL inner.removeEventListener.constructor.isInner should be true. Was false.
FAIL inner.resizeBy.isInner should be true. Was false.
FAIL inner.resizeBy.constructor.isInner should be true. Was false.
FAIL inner.resizeTo.isInner should be true. Was false.
FAIL inner.resizeTo.constructor.isInner should be true. Was false.
PASS inner.screen.isInner is true
PASS inner.screen.constructor.isInner is true
FAIL inner.screenLeft.isInner should be true. Was false.
FAIL inner.screenLeft.constructor.isInner should be true. Was false.
FAIL inner.screenTop.isInner should be true. Was false.
FAIL inner.screenTop.constructor.isInner should be true. Was false.
FAIL inner.screenX.isInner should be true. Was false.
FAIL inner.screenX.constructor.isInner should be true. Was false.
FAIL inner.screenY.isInner should be true. Was false.
FAIL inner.screenY.constructor.isInner should be true. Was false.
FAIL inner.scroll.isInner should be true. Was false.
FAIL inner.scroll.constructor.isInner should be true. Was false.
FAIL inner.scrollBy.isInner should be true. Was false.
FAIL inner.scrollBy.constructor.isInner should be true. Was false.
FAIL inner.scrollTo.isInner should be true. Was false.
FAIL inner.scrollTo.constructor.isInner should be true. Was false.
FAIL inner.scrollX.isInner should be true. Was false.
FAIL inner.scrollX.constructor.isInner should be true. Was false.
FAIL inner.scrollY.isInner should be true. Was false.
FAIL inner.scrollY.constructor.isInner should be true. Was false.
PASS inner.scrollbars.isInner is true
PASS inner.scrollbars.constructor.isInner is true
PASS inner.self.isInner is true
PASS inner.self.constructor.isInner is true
PASS inner.sessionStorage.isInner is true
PASS inner.sessionStorage.constructor.isInner is true
FAIL inner.setInterval.isInner should be true. Was false.
FAIL inner.setInterval.constructor.isInner should be true. Was false.
FAIL inner.setTimeout.isInner should be true. Was false.
FAIL inner.setTimeout.constructor.isInner should be true. Was false.
FAIL inner.status.isInner should be true. Was false.
FAIL inner.status.constructor.isInner should be true. Was false.
PASS inner.statusbar.isInner is true
PASS inner.statusbar.constructor.isInner is true
FAIL inner.stop.isInner should be true. Was false.
FAIL inner.stop.constructor.isInner should be true. Was false.
PASS inner.toolbar.isInner is true
PASS inner.toolbar.constructor.isInner is true
FAIL inner.webkitConvertPointFromNodeToPage.isInner should be true. Was false.
FAIL inner.webkitConvertPointFromNodeToPage.constructor.isInner should be true. Was false.
FAIL inner.webkitConvertPointFromPageToNode.isInner should be true. Was false.
FAIL inner.webkitConvertPointFromPageToNode.constructor.isInner should be true. Was false.
PASS inner.window.isInner is true
PASS inner.window.constructor.isInner is true
PASS successfullyParsed is true

TEST COMPLETE
